rag-chatbot/scripts/requirements.txt
google-generativeai>=0.4.1
qdrant-client==1.9.1
python-dotenv==1.0.0
PyYAML==6.0.1
markdown==3.4.4
beautifulsoup4==4.12.2